% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_level_features.R
\name{base_level_features}
\alias{base_level_features}
\title{Title}
\usage{
base_level_features(x, window_sizes = c(1, 4, 12), kick = NULL)
}
\arguments{
\item{kick}{}
}
\value{

}
\description{
Title
}
