% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_features.R
\name{basic_features}
\alias{basic_features}
\title{Title}
\usage{
basic_features(
  x,
  stat = c("sd", "mean", "median", "min", "max", "lm"),
  seasons = F,
  window = NULL
)
}
\arguments{
\item{window}{}
}
\value{

}
\description{
Title
}
