% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponentially_weighted_seasoons.R
\name{exponentially_weighted_seasons}
\alias{exponentially_weighted_seasons}
\title{Title}
\usage{
exponentially_weighted_seasons(x, window)
}
\arguments{
\item{window}{}
}
\value{

}
\description{
Title
}
