% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_tree_format.R
\name{transform_to_tree_format}
\alias{transform_to_tree_format}
\title{Title}
\usage{
transform_to_tree_format(x, h, var, keep = NULL)
}
\arguments{
\item{keep}{}
}
\value{

}
\description{
Title
}
